requirements.txt
asttokens==2.0.8
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.0.0
fonttools==4.37.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
ipykernel==6.15.2
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
jupyter_client==7.3.5
jupyter_core==4.11.1
kiwisolver==1.3.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
nest_asyncio==1.5.5
numpy==1.23.0
packaging==21.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
Pillow_SIMD==9.0.0.post1
promise==2.3
prompt_toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure_eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==23.1.0
requests==2.28.1
scikit_learn==1.1.2
scipy==1.8.0
sentry-sdk==1.9.5
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
stack-data==0.5.0
threadpoolctl==3.1.0
torch==1.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.2
wcwidth==0.2.5
